+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/State/Query.hs
@@ -1,13 +1,10 @@
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
-{-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -Wno-deprecations #-}
 
 module Cardano.Ledger.Api.State.Query (
@@ -37,15 +34,14 @@ module Cardano.Ledger.Api.State.Query (
   mkQueryPoolStateResult,
 
   -- * @GetStakeSnapshots@
-  queryStakeSnapshots,
-  StakeSnapshot (..),
-  StakeSnapshots (..),
+  module Cardano.Ledger.Api.State.Query.Snapshot,
 ) where
 
 import Cardano.Ledger.Api.State.Query.Epoch
 import Cardano.Ledger.Api.State.Query.Governance
 import Cardano.Ledger.Api.State.Query.PParams
-import Cardano.Ledger.BaseTypes (EpochNo, Network, NonZero, ProtVer (..))
+import Cardano.Ledger.Api.State.Query.Snapshot
+import Cardano.Ledger.BaseTypes (EpochNo, Network)
 import Cardano.Ledger.Binary
 import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Compactible (fromCompact)
@@ -59,14 +55,9 @@ import Cardano.Ledger.Conway.State
 import Cardano.Ledger.Core
 import Cardano.Ledger.Credential (Credential (..))
 import Cardano.Ledger.Shelley.LedgerState
-import Control.DeepSeq
-import Control.Monad (guard)
-import Data.Foldable (fold)
 import Data.Map (Map)
 import qualified Data.Map.Strict as Map
 import Data.Set (Set)
-import qualified Data.VMap as VMap
-import GHC.Generics
 import Lens.Micro
 
 -- | Implementation for @GetFilteredDelegationsAndRewardAccounts@ query.
@@ -174,141 +165,3 @@ queryPoolParameters ::
 queryPoolParameters network nes poolKeys =
   let pools = nes ^. nesEsL . esLStateL . lsCertStateL . certPStateL . psStakePoolsL
    in Map.mapWithKey (stakePoolStateToStakePoolParams network) $ Map.restrictKeys pools poolKeys
-
--- | The stake snapshot returns information about the mark, set, go ledger snapshots for a pool,
--- plus the total active stake for each snapshot that can be used in a 'sigma' calculation.
---
--- Each snapshot is taken at the end of a different era. The go snapshot is the current one and
--- was taken two epochs earlier, set was taken one epoch ago, and mark was taken immediately
--- before the start of the current epoch.
-data StakeSnapshot = StakeSnapshot
-  { ssMarkPool :: !Coin
-  , ssSetPool :: !Coin
-  , ssGoPool :: !Coin
-  }
-  deriving (Eq, Show, Generic)
-
-instance NFData StakeSnapshot
-
-instance EncCBOR StakeSnapshot where
-  encCBOR
-    StakeSnapshot
-      { ssMarkPool
-      , ssSetPool
-      , ssGoPool
-      } =
-      encodeListLen 3
-        <> encCBOR ssMarkPool
-        <> encCBOR ssSetPool
-        <> encCBOR ssGoPool
-
-instance DecCBOR StakeSnapshot where
-  decCBOR = do
-    enforceSize "StakeSnapshot" 3
-    StakeSnapshot
-      <$> decCBOR
-      <*> decCBOR
-      <*> decCBOR
-
-data StakeSnapshots = StakeSnapshots
-  { ssStakeSnapshots :: !(Map (KeyHash StakePool) StakeSnapshot)
-  , ssMarkTotal :: !(NonZero Coin)
-  , ssSetTotal :: !(NonZero Coin)
-  , ssGoTotal :: !(NonZero Coin)
-  }
-  deriving (Eq, Show, Generic)
-
-instance NFData StakeSnapshots
-
-instance EncCBOR StakeSnapshots where
-  encCBOR
-    StakeSnapshots
-      { ssStakeSnapshots
-      , ssMarkTotal
-      , ssSetTotal
-      , ssGoTotal
-      } =
-      encodeListLen 4
-        <> encCBOR ssStakeSnapshots
-        <> encCBOR ssMarkTotal
-        <> encCBOR ssSetTotal
-        <> encCBOR ssGoTotal
-
-instance DecCBOR StakeSnapshots where
-  decCBOR = do
-    enforceSize "StakeSnapshots" 4
-    StakeSnapshots
-      <$> decCBOR
-      <*> decCBOR
-      <*> decCBOR
-      <*> decCBOR
-
--- | Report stake per pool per snapshot as well as total active stake per snapshot.
---
--- /Note/ - Whenever poolIds are not supplied, we collect all of the pools, even if they don't have
--- any delegations. Otherwise we filter out for exact poolIds that were supplied. In both cases it
--- means that there can be pools that have zero stake in all three snapshot, but the meaning of that
--- can be very different:
---
--- * either the pool has no delegations, or
--- * it was explicitly requested even though it has no stake or not even registered
---
--- However, starting with Protocol Version 11 we remove this strange inconsistency and only ever
--- report stake pools with non-zero stake, which means pools without delegations (hence without any stake in any of the three snapshots) are no longer included in the results.
-queryStakeSnapshots ::
-  EraGov era =>
-  NewEpochState era ->
-  Maybe (Set (KeyHash StakePool)) ->
-  StakeSnapshots
-queryStakeSnapshots nes mPoolIds =
-  let SnapShots
-        { ssStakeMark
-        , ssStakeSet
-        , ssStakeGo
-        } = esSnapshots $ nesEs nes
-
-      mkStakeSnapshotMaybe poolId = do
-        let
-          markPoolStake = spssStake <$> VMap.lookup poolId (ssStakePoolsSnapShot ssStakeMark)
-          setPoolStake = spssStake <$> VMap.lookup poolId (ssStakePoolsSnapShot ssStakeSet)
-          goPoolStake = spssStake <$> VMap.lookup poolId (ssStakePoolsSnapShot ssStakeGo)
-        -- Non-registered stake pools or ones that have no stake are of no interest to us.
-        guard (fold [markPoolStake, setPoolStake, goPoolStake] > Just mempty)
-        Just
-          StakeSnapshot
-            { ssMarkPool = maybe mempty fromCompact markPoolStake
-            , ssSetPool = maybe mempty fromCompact setPoolStake
-            , ssGoPool = maybe mempty fromCompact goPoolStake
-            }
-      mkStakeSnapshot poolId =
-        let
-          lookupStake =
-            maybe mempty (fromCompact . spssStake) . VMap.lookup poolId . ssStakePoolsSnapShot
-         in
-          StakeSnapshot
-            { ssMarkPool = lookupStake ssStakeMark
-            , ssSetPool = lookupStake ssStakeSet
-            , ssGoPool = lookupStake ssStakeGo
-            }
-      version = pvMajor (nes ^. nesEsL . curPParamsEpochStateL . ppProtocolVersionL)
-      poolIds =
-        case mPoolIds of
-          Nothing
-            | version < natVersion @11 ->
-                foldMap
-                  (VMap.keysSet . VMap.filter (\_ -> (> 0) . spssNumDelegators) . ssStakePoolsSnapShot)
-                  [ssStakeMark, ssStakeSet, ssStakeGo]
-            | otherwise ->
-                foldMap
-                  (VMap.keysSet . VMap.filter (\_ -> (> mempty) . spssStake) . ssStakePoolsSnapShot)
-                  [ssStakeMark, ssStakeSet, ssStakeGo]
-          Just ids -> ids
-   in StakeSnapshots
-        { ssStakeSnapshots =
-            if version < natVersion @11
-              then Map.fromSet mkStakeSnapshot poolIds
-              else Map.mapMaybe id $ Map.fromSet mkStakeSnapshotMaybe poolIds
-        , ssMarkTotal = ssTotalActiveStake ssStakeMark
-        , ssSetTotal = ssTotalActiveStake ssStakeSet
-        , ssGoTotal = ssTotalActiveStake ssStakeGo
-        }