+++ libs/cardano-ledger-api/testlib/Test/Cardano/Ledger/Api/Arbitrary.hs
@@ -34,10 +34,10 @@ instance Arbitrary QueryResultDRepStates where
 instance Arbitrary QueryPoolStateResult where
   arbitrary = QueryPoolStateResult <$> arbitrary <*> arbitrary <*> arbitrary <*> arbitrary
 
-instance Arbitrary StakeSnapshot where
+instance Arbitrary QueryResultStakeSnapshot where
   arbitrary = genericArbitraryU
   shrink = genericShrink
 
-instance Arbitrary StakeSnapshots where
+instance Arbitrary QueryResultStakeSnapshots where
   arbitrary = genericArbitraryU
   shrink = genericShrink